+++ docker/hbase-1.1.2/build-image.sh
@@ -1,0 +1,1 @@
+docker build --rm=true -t nerdammer/hbase:1.1.2 .