+++ docker/hbase-1.1.2/start-pseudo-distributed.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+/opt/hbase/bin/hbase zookeeper > logzoo.log 2>&1 &
+/opt/hbase/bin/hbase regionserver start > logregion.log 2>&1 &
+
+/opt/hbase/bin/hbase master start --localRegionServers=0