+++ docker/hbase-rest/start-rest.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+/opt/hbase/bin/start-pseudo-distributed.sh > logmaster.log 2>&1 &
+
+/opt/hbase/bin/hbase rest start
+